Attempt to do Ajax with Haste's methods.
Use Haste's jsonRequest function to perform an AJAX call. This
*almost* works. It fetches the right data from the server, and
sets $scope.phones to the proper array. But the view never seems
to notice that the model has gotten data.

I suspect there's some function we can call to get the view to
refresh, so this approach can probably be made to work with a bit
more understanding of Angular's internals.